+++ port_ocean/tests/core/utils/test_entity_topological_sorter.py
@@ -1,5 +1,5 @@
 from port_ocean.core.models import Entity
-from port_ocean.utils.entity_topological_sorter import EntityTopologicalSorter
+from port_ocean.core.utils.entity_topological_sorter import EntityTopologicalSorter
 from unittest.mock import MagicMock
 from port_ocean.exceptions.core import (
     OceanAbortException,